Virtual Key
As a Full Stack Developer, complete the features of the application by planning the development in terms of sprints and then push the source code to the GitHub repository. As this is a prototyped application, the user interaction will be via a command line. 

The flow and features of the application:

Plan more than two sprints to complete the application

Document the flow of the application and prepare a flow chart 

List the core concepts and algorithms being used to complete this application

Code to display the welcome screen. It should display:

Application name and the developer details 

The details of the user interface such as options displaying the user interaction information 

Features to accept the user input to select one of the options listed 

The first option should return the current file names in ascending order. The root directory can be either empty or contain few files or folders in it

 The second option should return the details of the user interface such as options displaying the following:

Add a file to the existing directory list

You can ignore the case sensitivity of the file names 

Delete a user specified file from the existing directory list

You can add the case sensitivity on the file name in order to ensure that the right file is deleted from the directory list

Return a message if FNF (File not found)

Search a user specified file from the main directory

You can add the case sensitivity on the file name to retrieve the correct file

Display the result upon successful operation

Display the result upon unsuccessful operation

Option to navigate back to the main context

There should be a third option to close the application

Implement the appropriate concepts such as exceptions, collections, and sorting techniques for source code optimization and increased performance